Move landing page component. Add intro info.
